requirements.txt
Django
djangorestframework
djangorestframework-simplejwt   
celery
redis
openpyxl
drf-yasg
requests
